requirements.txt
Pillow
streamlit
boto3

